refactor package and test names
